pyproject.toml
[project]
name = "django-post-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.8",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "djangorestframework>=3.16.1",
    "drf-yasg>=1.21.11",
]

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]
